Move redirect after submit logic to the review page.
